refactor: extract shared native foundation for windows/linux
