+++ engines/linux-native/Cargo.toml
@@ -4,6 +4,4 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-serde_json = "1"
-protocol-rust = { path = "../protocol-rust" }
-time = { version = "0.3", features = ["formatting"] }
+native-foundation = { path = "../native-foundation" }
